requirements.txt
fastapi
uvicorn
python-dotenv
stytch
supabase
langchain-openai
